Create minstack.js
MinStack is a data structure that supports push, pop, top, and retrieval of the minimum element in constant time. It maintains two stacks internally: one for the actual stack elements and one for tracking the minimum elements.

Features
Push: Add an element to the stack.
Pop: Remove the top element from the stack.
Top: Retrieve the top element of the stack.
GetMin: Retrieve the minimum element in the stack.
Usage
Installation
No installation is required; you can simply include the MinStack implementation in your JavaScript project.

Methods
MinStack()
Description: Initializes a new instance of the MinStack.
push(val)
Parameters: val (number) - The value to be added to the stack.
Description: Adds val to the stack and updates the minimum stack.
pop()
Description: Removes the top element from the stack and the minimum stack.
Throws: Throws an error if the stack is empty.
top()
Returns: The top element of the stack.
Description: Retrieves the top element of the stack.
Throws: Throws an error if the stack is empty.
getMin()
Returns: The minimum element of the stack.
Description: Retrieves the minimum element in the stack.
Throws: Throws an error if the stack is empty.
Error Handling
Empty Stack: Throws an error with the message 'the stack is empty' if an operation is performed on an empty stack.
License
This project is licensed under the MIT License. See the LICENSE file for details.